removed seperate import tests for VMSS
